style: prevent file too large
